refactor(MetadataMapper): move to adapters folder
