+++ src/get-saml-metadata/src/application/adapters/MetadataMapperAdapter.ts
